Implement new render strategy
The grid is now procedurally generated in the shader and the updated
values are passed as a uniform height-map texture.
The previous implementation used transient buffers which were not made
for the job as they are much restricted in size. (max 2 MiB for tib)
This new implementation allows grid sizes as large as the local RAM
allows. However grids with millions of cells are too large for smooth
animations. An approach using compute shaders is planned for the future.